requirements.txt
customtkinter
numpy
opencv-python
Pillow